+++ src/enums/process-attribute.enum.ts
@@ -1,0 +1,5 @@
+export enum ProcessAttribute {
+  COMMAND = "process.command",
+  COMMAND_ARGS = "process.command_args",
+  OWNER = "process.owner",
+}